+++ packages/import-maps-generate/test/generateFromYarnLock.test.js
@@ -4,7 +4,7 @@ import {
   generateFromYarnLock,
   resolvePathsAndConflicts,
   flatResolvedDepsToImports,
-} from '../src/generateFromYarnLock';
+} from '../src/generateFromYarnLock.js';
 
 const { expect } = chai;