+++ packages/create/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.9](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.12.8...@open-wc/create@0.12.9) (2019-07-08)
+
+
+### Bug Fixes
+
+* use file extensions for imports to support import maps ([c711b13](https://github.com/open-wc/open-wc/commit/c711b13))
+
+
+
+
+
 ## [0.12.8](https://github.com/open-wc/open-wc/compare/@open-wc/create@0.12.7...@open-wc/create@0.12.8) (2019-06-23)