+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/open-wc/open-wc/compare/@open-wc/testing@1.0.9...@open-wc/testing@1.0.10) (2019-07-08)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
+## [1.0.9](https://github.com/open-wc/open-wc/compare/@open-wc/testing@1.0.8...@open-wc/testing@1.0.9) (2019-07-08)
+
+**Note:** Version bump only for package @open-wc/testing
+
+
+
+
+
 ## [1.0.8](https://github.com/open-wc/open-wc/compare/@open-wc/testing@1.0.7...@open-wc/testing@1.0.8) (2019-07-08)
 
 **Note:** Version bump only for package @open-wc/testing